pyproject.toml
[project]
name = "crawler-manual"
version = "0.1.0"
description = ""
authors = [
    {name = "Do-hyun Ko (Dori)",email = "dohyun682@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "numpy (>=2.4.1,<3.0.0)",
    "google-adk (>=1.22.0,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false